Move route tests out of index.test.js
